+++ aten/src/ATen/core/TensorMethods.h
@@ -2,6 +2,7 @@
 
 #include "ATen/core/Tensor.h"
 #include <c10/core/Scalar.h>
+#include <c10/macros/Macros.h>
 #include "ATen/core/SparseTensorRef.h"
 #include "ATen/core/Type.h"
 #include "ATen/core/TensorOptions.h"
