+++ c10/core/TensorOptions.cpp
@@ -1,4 +1,4 @@
-#include <ATen/core/TensorOptions.h>
+#include <c10/core/TensorOptions.h>
 
 #include <c10/Device.h>
 #include <c10/core/Layout.h>
@@ -7,7 +7,7 @@
 
 #include <iostream>
 
-namespace at {
+namespace c10 {
 
 std::ostream& operator<<(
     std::ostream& stream,
@@ -19,4 +19,4 @@ std::ostream& operator<<(
                 << options.requires_grad() << ")";
 }
 
-} // namespace at
+} // namespace c10
